Feat: UICollectionViewCell 커스텀 클래스 구현
